Bug Fix
Fixed the following bug: A file containing multiple classes, or inheritant classes createad a bug. All subsequent Get-Cu* cmdlets where not displaying doublons
For example, get-culoadedclass, for a file containing 3 classes, was returning 3 times a RAW ast instead of one.
I changed the way get-cuclass was creating a new class. First we create the class, then check if the AST contains more than one entry. If that's the case, it means there are more than one class in the class file.