Bug Fix in out-cupsgraph + rewrote of write-cuclassdiagram
write-classdiagram has a new parameter outputdiagram (doesnt mind if we change it's name...) to determine if the output file are file(s) based, or directory based.

Bug fix in out-cupsgraph: when a property was referenced in an other class, and the property was an array, the composistion did not work. Problem fixed with a regex replace... i alsod added some verbose output... helped me solve the stuff... !